+++ smarttime/src/main/java/com/smarttime/repository/BusinessRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by michal.berentowicz on 14.05.2017.
  */
-public interface BusinessRepository extends JpaRepository<Business, Long> {
+public interface BusinessRepository extends JpaRepository<Business, Integer> {
 }
